ocr_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai>=1.52.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
